requirements.txt
alabaster==0.7.10
appdirs==1.4.3
Babel==2.4.0
blessings==1.6
certifi==2017.4.17
chardet==3.0.4
colorama==0.3.9
docutils==0.13.1
enum34==1.1.6
gitdb2==2.0.2
GitPython==2.1.5
idna==2.5
imagesize==0.7.1
invocations==0.19.0
invoke==0.23.0
Jinja2==2.9.6
MarkupSafe==1.0
packaging==16.8
Pygments==2.2.0
pyparsing==2.2.0
pytz==2017.2
releases==1.3.1
requests==2.18.1
semantic-version==2.6.0
six==1.10.0
smmap2==2.0.3
snowballstemmer==1.2.1
Sphinx==1.4.9
tabulate==0.7.5
tqdm==4.14.0
urllib3==1.21.1
alabaster==0.7.10
Babel==2.5.3
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.9
docutils==0.14
gitdb2==2.0.3
GitPython==2.1.9
idna==2.6
imagesize==1.0.0
invoke==0.23.0
Jinja2==2.10
MarkupSafe==1.0
packaging==17.1
Pygments==2.2.0
pyparsing==2.2.0
pytz==2018.4
requests==2.18.4
six==1.11.0
smmap2==2.0.3
snowballstemmer==1.2.1
Sphinx==1.7.4
sphinxcontrib-websupport==1.0.1
urllib3==1.22
virtualenv==15.2.0
